PANA-7720 Move image redaction onto snapshots
